+++ pycheops/__init__.py
@@ -29,5 +29,5 @@
 try:
           from .version import __version__
 except:
-          __version__ = '0.1.8'
+          __version__ = '0.1.9'